+++ src/DTF_ESP32Update.h
@@ -1,34 +1,32 @@
 #ifndef DTF_ESP32Update_H
 #define DTF_ESP32Update_H
 
-enum class DTF_UpdateResponse
+namespace DTF_ESP32Update
 {
-    UPDATE_FAILED = -1,
-    NO_UPDATE_AVAILABLE = 1,
-    UPDATE_SUCCEEDED = 2,
+
+enum class UpdateResponse
+{
+    UnknownError = -1,
+    NewFirmwareFlashed = 0,
+    InvalidFirmwareImage = 1,
+    NotEnoughMemory = 2,
+    FirmwareWriteFailed = 3,
+    NoUpdatesAvailable = 4,
+    CertValidationFailed = 5,
 };
 
-enum class DTF_SetTimeOption : int
-{
-    DO_NOT_SET_TIME = 0,
-    SET_TIME = 1,
-};
-
-enum class DTF_RebootOption : int
+enum class RebootOption : int
 {
     NO_REBOOT = 0,
     REBOOT = 1,
 };
 
-class DTF_ESP32Update
-{
-    public:
-        static DTF_UpdateResponse getFirmwareUpdate(
-            const char* updateUrl, 
-            const char* currentVersion, 
-            DTF_RebootOption rebootOption = DTF_RebootOption::REBOOT, 
-            DTF_SetTimeOption setTimeOption = DTF_SetTimeOption::DO_NOT_SET_TIME
+UpdateResponse getFirmwareUpdate(
+            const char* productId, 
+            const char* currentVersion,
+            RebootOption rebootOption = RebootOption::REBOOT
         );   
-};
+
+}  // namespace DTF_Arduino
 
 #endif // DTF_ESP32Update_H